+++ showlinks.py
@@ -1,4 +1,4 @@
-#!/Users/pberry/src/ridehome/env/bin/python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-